Refactor match logic and fix abort logic bugs
